# docs/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: Coeeter
    url: https://github.com/coeeter/aniways
  description: An anime tracking and streaming platform API
  title: AniWays API
  version: 1.0.0
paths:
  "/anime/{id}":
    get:
      description: Get anime by ID
      parameters:
        - description: Anime ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.AnimeWithMetadataResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Get anime by ID
      tags:
        - Anime
  "/anime/{id}/banner":
    get:
      description: Get anime banner
      parameters:
        - description: Anime ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.BannerResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Get anime banner
      tags:
        - Anime
...
  "/anime/{id}/franchise":
    get:
      description: Get anime franchise relations
      parameters:
        - description: Anime ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.RelationsResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Get anime franchise relations
      tags:
        - Anime
  "/anime/{id}/trailer":
    get:
      description: Get anime trailer
      parameters:
        - description: Anime ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.TrailerResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Get anime trailer
      tags:
        - Anime
...
servers:
  - url: http://localhost:8080
components:
...
  securitySchemes:
    cookieAuth:
      in: cookie
      name: aniways_session
      type: apiKey
  schemas:
...
    models.AnimeMetadataResponse:
      properties:
        airingEndDate:
          example: 2023-09-30
          type: string
        airingStartDate:
          example: 2023-04-01
          type: string
        airingStatus:
          example: finished_airing
          type: string
        avgEpisodeDuration:
          example: 24
          type: integer
        description:
          example: Humanity fights for survival against giant humanoid Titans.
          type: string
        mainPictureUrl:
          example: https://example.com/main.jpg
          type: string
        malId:
          example: 12345
          type: integer
        mean:
          example: 9
          type: number
        mediaType:
          example: TV
          type: string
        popularity:
          example: 1
          type: integer
        rank:
          example: 5
          type: integer
        rating:
          example: R - 17+ (violence & profanity)
          type: string
        scoringUsers:
          example: 500000
          type: integer
        season:
          example: spring
          type: string
        seasonYear:
          example: 2023
          type: integer
        source:
          example: Manga
          type: string
        studio:
          example: Studio Pierrot
          type: string
        totalEpisodes:
          example: 25
          type: integer
        trailerEmbedUrl:
          example: https://www.youtube.com/embed/abc123
          type: string
      type: object
    models.AnimeResponse:
      properties:
        anilistId:
          example: 67890
          type: integer
        ename:
          example: Attack on Titan
          type: string
        genre:
          example: Action, Drama
          type: string
        id:
          example: V1StGXR8Z5jdHi6B
          type: string
        imageUrl:
          example: https://example.com/anime/image.jpg
          type: string
        jname:
          example: 進撃の巨人
          type: string
        lastEpisode:
          example: 25
          type: integer
        malId:
          example: 12345
          type: integer
        season:
          example: spring
          type: string
        seasonYear:
          example: 2023
          type: integer
      type: object
    models.AnimeWithMetadataResponse:
      properties:
        anilistId:
          example: 67890
          type: integer
        ename:
          example: Attack on Titan
          type: string
        genre:
          example: Action, Drama
          type: string
        id:
          example: V1StGXR8Z5jdHi6B
          type: string
        imageUrl:
          example: https://example.com/anime/image.jpg
          type: string
        jname:
          example: 進撃の巨人
          type: string
        lastEpisode:
          example: 25
          type: integer
        malId:
          example: 12345
          type: integer
        metadata:
          $ref: "#/components/schemas/models.AnimeMetadataResponse"
        season:
          example: spring
          type: string
        seasonYear:
          example: 2023
          type: integer
      type: object
    models.BannerResponse:
      properties:
        url:
          example: https://example.com/banner.jpg
          type: string
      type: object
...
    models.ErrorResponse:
      properties:
        error:
          example: Invalid request
          type: string
      type: object
...
    models.RelationsResponse:
      properties:
        related:
          items:
            $ref: "#/components/schemas/models.AnimeResponse"
          type: array
        watchOrder:
          items:
            $ref: "#/components/schemas/models.AnimeResponse"
          type: array
      type: object
...
    models.TrailerResponse:
      properties:
        trailer:
          example: https://www.youtube.com/embed/abc123
          type: string
      type: object
...